+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## [vNext]
+### Added
+-  The `USE_PREDEFINED_DEVICE_INFO` feature now use viewport height and status bar height from server. [Trello 2307](https://trello.com/c/8VCtSmfN)
+
 ## [3.193.0]
 ### Added
 - Supporting OCR Feature. [Trello 2440](https://trello.com/c/FLGKnqIS)